requirements.txt
Django>=5.1.0
djangorestframework>=3.14.0
django-cors-headers>=4.3.1
django-cleanup>=8.1.0
Pillow>=10.2.0 